+++ kotlin-collections-projeto-inicial/src/br/com/alura/main.kt
@@ -1,106 +1,23 @@
 package br.com.alura
 
 fun main() {
-    val pedidos: MutableMap<Int, Double> = mutableMapOf(
-        Pair(1, 20.0),
-        Pair(2, 64.0),
-        3 to 50.0,
-        4 to 150.0,
-        5 to 150.0,
-        6 to 80.0
+    val pedidos = listOf(
+        Pedido(1, 20.0),
+        Pedido(2, 60.0),
+        Pedido(3, 30.0),
+        Pedido(4, 70.0)
     )
+    println(pedidos)
+    val pedidosMapeados: Map<Int, Pedido> = pedidos.associateBy { pedido -> pedido.numero }
 
-//    val valorPedido = pedidos.getValue(5)
-//    println(valorPedido)
+    println(pedidosMapeados)
+    println(pedidosMapeados[1])
 
-    val mensagem: Double = pedidos.getOrElse(0) {
-        0.0
+    val pedidosFreteGratis: Map<Pedido, Boolean> = pedidos.associateWith { pedido ->
+        pedido.valor > 50.0
     }
-    println(mensagem)
-    println(pedidos.getOrDefault(0, -1.0))
-
-    println(pedidos.keys)
-
-    for (numero in pedidos.keys) {
-        println("Pedido $numero")
-    }
-
-    println(pedidos.values)
-
-    for (valor in pedidos.values) {
-        println("Valor $valor")
-    }
-
-    val pedidosFiltrados = pedidos.filter { (numero, valor) ->
-        numero % 2 == 0 && valor > 50.0
-    }
-    println(pedidosFiltrados)
-
-    val pedidosAcima = pedidos.filterValues { valor ->
-        valor > 70.0
-    }
-    println(pedidosAcima)
-
-    val pedidosPares = pedidos.filterKeys { numero ->
-        numero % 2 == 0
-    }
-    println(pedidosPares)
-
-//    println(pedidos + Pair(7, 90.0))
-    println(pedidos + mapOf(7 to 90.0, 8 to 20.0))
-    println(pedidos)
-
-    println(pedidos - listOf(6, 5))
-    println(pedidos)
-
-//    pedidos.putAll(setOf<Pair<Int, Double>>(7 to 90.0, 8 to 20.0))
-    pedidos += setOf<Pair<Int, Double>>(7 to 90.0, 8 to 20.0, 8 to 30.0)
-    println(pedidos)
-
-    pedidos.keys.remove(1)
-    println(pedidos)
-
-    pedidos.values.remove(50.0)
-    println(pedidos)
-
-    pedidos.values.remove(150.0) //remove o primeiro elemento que encontra com o valor passado
-    println(pedidos)
-
-    pedidos -= 6
-    println(pedidos)
+    println(pedidosFreteGratis)
+    println(pedidosFreteGratis[Pedido(numero=2, valor=60.0)])
 }
 
-fun testaMap(pedidos: MutableMap<Int, Double>) {
-    println(pedidos)
-
-    val pedido = pedidos[1]
-    pedido?.let {
-        println("pedido $it")
-    }
-
-    for (p: Map.Entry<Int, Double> in pedidos) {
-        println("Número do pedido: ${p.key}")
-        println("Valor do pedido: ${p.value}")
-    }
-
-    pedidos[4] = 70.0
-    println(pedidos)
-
-    pedidos.put(5, 80.0)
-    println(pedidos)
-
-    pedidos[1] = 100.0
-    println(pedidos)
-
-    pedidos.putIfAbsent(6, 200.0)
-    println(pedidos)
-
-    pedidos.putIfAbsent(6, 300.0)
-    println(pedidos)
-
-    pedidos.remove(6)
-    println(pedidos)
-
-    pedidos.remove(3, 50.0)
-    println(pedidos)
-}
+data class Pedido(val numero: Int, val valor: Double)
